Don't barf on Yaml with tags
Yaml can contain custom tags in the form of '!vault' and such. Before
this commit, this would cause Yaml to be unable to read the entire yaml
file.

This commit introduces a custom yaml loader that implements the
SafeLoader class and overrides what happens when tags are encountered.
For `vault` tags, it returns an 'ENCRYPTED CONTENTS REDACTED' string.
For all other (unknown) tags, it returns the tag value as a string.

This should also fix "Skipping encrypted file'-like warnings when
generating CMDBs.