use python version define in PY_BIN if it exists
Check if we've an executable file provided in an env var called
`PY_BIN`. If that's the case make it the first choice for the
validation of the python version requirements.
As a fallback populate the list of python options to try with
all `python` and `python3` versions we can find in the `PATH`
of the environment.