Allow 'ansible-cmdb' wrapper to be a symlink
Enhances 'ansible-cmdb' wrapper script to be able to find 'ansible-cmdb.py' when 'ansible-cmdb' is a symlink.

Currently the search paths around 'ansible-cmdb' are relative to the location of the symlink. We can use 'readlink -f'
to instead follow the symlink and search from the actual location of 'ansible-cmdb'

The use-case where this comes up is we want to add 'ansible-cmdb' to the path. To do this we are creating a symlink
from '/bin/ansible-cmdb' to its actual location. After this update,  the command '/bin/ansible-cmdb' works as expected
because the wrapper script follows the symlink.