Fix python binary lookup on debian bullseye (11)
On debian bullseye, the `/usr/bin/python` symlink is [intentionally removed](https://wiki.debian.org/Python#Python_in_Debian) (see the note). With this change, the wrapper detects `python3`, `python2` and `python` symlinks in this order.

The error that lead me to this was (added so that searches on github will find this PR):
`No suitable python version found (v2.7 or higher required). Aborting`